requirements.txt
nextcord==2.6.0
python-dotenv==1.0.0
pyfiglet==0.8.post1
termcolor==2.3.0
requests==2.31.0
youtube-dl==2021.12.17
yt-dlp==2023.12.30
Pillow==10.1.0
aiohttp==3.9.1
asyncio==3.4.3
rich==13.7.0
colorama==0.4.6
tqdm==4.66.1 